requirements.txt
langchain
langchain-google-genai
chromadb
sentence_transformers
langchain_huggingface
langchain_chroma
python-dotenv
langchain_community
ipykernel
sentence-transformers
langchain_core
streamlit
pdfminer.six
python-docx